Centralise ANSI stripping helpers
Move the escape removal logic into the shared test utilities so both
unit and integration tests reuse it, simplify the CLI regression's diff
checks, and adopt char literals for the HTML normaliser constants.
